+++ Node-Beaver/Node-Beaver.cydsn/usb_manager.c
@@ -17,8 +17,8 @@ void usb_init()
 
 
 
-void usb_put(const DataPacket* data_queue, uint8_t data_head,
-	uint8_t data_tail)
+void usb_put(const DataPacket* data_queue, uint16_t data_head,
+	uint16_t data_tail)
 {
 	uint8_t pos;
 	uint32_t num_char;
